Rename lambda functions to make their purposes clearer
