+++ .webpack/webpack.common.js
@@ -3,8 +3,9 @@ const ForkTsCheckerWebpackPlugin = require("fork-ts-checker-webpack-plugin")
 
 module.exports = {
   entry: {
-    compare: "./src/comparison/lambdas/compareLambda.ts",
-    rerunComparison: "./src/comparison/lambdas/rerunComparisonLambda.ts"
+    compareSingle: "./src/comparison/lambdas/compareSingle.ts",
+    gatherStats: "src/comparison/lambdas/gatherStats.ts",
+    rerunFailures: "./src/comparison/lambdas/rerunFailures.ts"
   },
   resolve: {
     modules: [path.resolve("./node_modules"), path.resolve("."), path.resolve(".")],
@@ -21,13 +22,12 @@ module.exports = {
         "bichard7-next-data": {
           test: /[\\/]node_modules[\\/]bichard7-next-data-.*?[\\/]/,
           name(module) {
-            const moduleFileName = module
-              .identifier()
-              .match(/(?<moduleName>bichard7-next-data-.*?)[\\/]/).groups.moduleName
-            return moduleFileName;
+            const moduleFileName = module.identifier().match(/(?<moduleName>bichard7-next-data-.*?)[\\/]/)
+              .groups.moduleName
+            return moduleFileName
           },
           chunks: "all"
-        },
+        }
       }
     }
   },
