+++ src/comparison/lambdas/compareSingle.integration.test.ts
@@ -6,7 +6,7 @@ import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3"
 import type { DocumentClient } from "aws-sdk/clients/dynamodb"
 import fs from "fs"
 import MockDate from "mockdate"
-import lambda from "src/comparison/lambdas/compareLambda"
+import lambda from "src/comparison/lambdas/compareSingle"
 import MockDynamo from "tests/helpers/MockDynamo"
 import MockS3 from "tests/helpers/MockS3"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
