+++ src/comparison/lambdas/compareSingle.ts
