+++ src/comparison/lambdas/gatherStats.ts
